Fix #288 - Delay video tag event listening
If we attach video handlers on same eventloop tick that we modify the
video tag, in first frame setups the video tag might in some cases emit
error events we are not interested in.

Now we let the load-method finish and the browser do it's things before
attaching video tag handlers.

This needs thorough testing with different setups and devices so that we
don't accidentally break anything.