dvr: display timestamp (#1137)
The timestamp is especially informative for DVR as it can show the
current or desired position inside the DVR window. So show it.

The CSS hiding of the timestamp for live (not dvr) can be omitted. It
was redundant anyway because is-live hides the timeline, and therefore
one cannot mousemove over it.